Complete test API findAll
+ Can get data from role table

+ Problem:
	+ Error when using hql
	+ HibernateUtil.java must modify to mapping with db when pass an entity to API

+++ toeic-core-persistence-data/src/main/java/vn/myclass/core/persistence/entity/RoleEntity.java
@@ -8,7 +8,7 @@ import java.util.List;
 public class RoleEntity {
     @Id
     @Column(name = "roleid")
-    private short roleId;
+    private Integer roleId;
 
     @Column(name = "name")
     private String name;
@@ -16,11 +16,11 @@ public class RoleEntity {
     @OneToMany(mappedBy = "roleEntity", fetch = FetchType.LAZY)
     private List<UserEntity> userEntityList;
 
-    public short getRoleId() {
+    public Integer getRoleId() {
         return roleId;
     }
 
-    public void setRoleId(short roleId) {
+    public void setRoleId(Integer roleId) {
         this.roleId = roleId;
     }
